+++ doc/source/user_guide/session/launching_ansys_fluent.rst
@@ -291,7 +291,7 @@ can be launched explicitly using :func:`create_launcher() <ansys.fluent.core.lau
 .. code:: python
 
   >>> from ansys.fluent.core.launcher.launcher import create_launcher
-  >>> from ansys.fluent.core.launcher.pyfluent_enums import LaunchMode, FluentMode
+  >>> from ansys.fluent.core.launcher.launch_options import LaunchMode, FluentMode
 
   >>> pim_meshing_launcher = create_launcher(LaunchMode.PIM, mode=FluentMode.MESHING)
   >>> pim_meshing_session = pim_meshing_launcher()
