+++ src/ansys/fluent/core/__init__.py
@@ -36,17 +36,17 @@ from ansys.fluent.core.get_build_details import (  # noqa: F401
     get_build_version,
     get_build_version_string,
 )
-from ansys.fluent.core.launcher.launcher import (  # noqa: F401
-    connect_to_fluent,
-    launch_fluent,
-)
-from ansys.fluent.core.launcher.pyfluent_enums import (  # noqa: F401
+from ansys.fluent.core.launcher.launch_options import (  # noqa: F401
     Dimension,
     FluentLinuxGraphicsDriver,
     FluentMode,
     FluentWindowsGraphicsDriver,
     Precision,
     UIMode,
+)
+from ansys.fluent.core.launcher.launcher import (  # noqa: F401
+    connect_to_fluent,
+    launch_fluent,
 )
 from ansys.fluent.core.parametric import LocalParametricStudy  # noqa: F401
 from ansys.fluent.core.pyfluent_warnings import (  # noqa: F401
