+++ src/ansys/fluent/core/launcher/__init__.py
@@ -27,7 +27,7 @@ from .fluent_container import (  # noqa: F401
     configure_container_dict,
     start_fluent_container,
 )
+from .launch_options import LaunchMode  # noqa: F401
 from .launcher import create_launcher  # noqa: F401
 from .pim_launcher import launch_remote_fluent  # noqa: F401
 from .process_launch_string import get_fluent_exe_path  # noqa: F401
-from .pyfluent_enums import LaunchMode  # noqa: F401
